Instructions: sync from riscv-opcodes
